[Uninstall Command] Complete uninstall command

+++ src/commands/uninstall.rs
@@ -13,13 +13,18 @@ pub fn uninstall() {
     remove_dir(&adof_dir);
 
     let output = Command::new("which").arg("adof").output().unwrap();
-    println!("output: {:?}", output);
     let adof_bin_dir = String::from_utf8(output.stdout).unwrap();
     remove_dir(adof_bin_dir.trim());
 }
 
 fn remove_dir(dir: &str) {
     let path = Path::new(dir);
+
+    if path.is_dir() {
     fs::remove_dir_all(path).unwrap();
-    println!("{:?}", dir);
+    }
+
+    if path.is_file() {
+        fs::remove_file(path).unwrap();
+    }
 }
